+++ src/main/java/org/project/wherego/member/controller/MemberController.java
@@ -6,6 +6,8 @@ import org.project.wherego.member.dto.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.project.wherego.member.service.MemberService;
@@ -49,4 +51,13 @@ public class MemberController {
         memberService.logout();
         return ResponseEntity.ok(new LogoutResponse("로그아웃 성공"));
     }
+
+    @GetMapping("/mypage")          // 로그인된 사용자 정보 받기
+    public ResponseEntity<?> mypageInfo(@AuthenticationPrincipal UserDetails userDetails) {
+        String email = userDetails.getUsername();
+        MyPageResponse mypageresponse = memberService.mypageInfo(email);
+        return ResponseEntity.ok(mypageresponse);
+    }
+
 }
+
