+++ src/main/java/org/project/wherego/member/dto/MyPageResponse.java
@@ -1,0 +1,16 @@
+package org.project.wherego.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class MyPageResponse {
+    private String email;
+    private String password;
+    private String nickname;
+}
